Integrate Firebase and add Analytics dependency
This commit integrates Firebase into the project.

It adds the `com.google.gms.google-services` plugin to both the project-level and app-level `build.gradle.kts` files. The Firebase Bill of Materials (BOM) is added to `gradle/libs.versions.toml` and implemented in the app's dependencies to manage Firebase library versions. Additionally, the `firebase-analytics` library is included.